ecommerce/requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
confluent-kafka==2.5.3
Django==4.2.16
django-cors-headers==4.4.0
django-environ==0.11.2
django-mongodb-engine==0.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangotoolbox==1.8.0
dnspython==2.6.1
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.4
graphql-relay==3.2.0
idna==3.10
kafka-python==2.0.2
mongoengine==0.29.1
pika==1.3.2
promise==2.3
psycopg2-binary==2.9.9
PyJWT==2.9.0
pymongo==4.9.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.3
